Revert "Revert "feat: Re-organized structure""
