--- src/domain/index.ts
@@ -1,2 +1,0 @@
-export * from "./errors/custom.error";
-export * from "./shared/index";
