--- src/domain/shared/index.ts
@@ -1,1 +1,0 @@
-export * from "./path.util";
